Move versions router to own file
